Beginning the long-awaited massive overhaul of ADAQAnalysisGUI. The giant ~7000 lines of code that form a single class and main() method will be split into many files...
  ADAQAnalysis.cc --> main()

  ADAQAnalysisInterface.cc/.hh --> ROOT GUI interface

  ADAQAnalysisManager.cc/.hh --> Custom data algorithms, file i/o, plotting, etc


This check contains the first refactoring of splitting the main method
from the class structures, which required a significant overhaul of
the Makefile to handle the build. ADAQAnalysisGUI class has been
renamed to ADAQAnalysisInterface, ADAQAnalysisManager class has been
added (it's blank for now), and the makefile updated to support full
binary and parallel builds in the new structure.