Summary: Implemented new features, mainly loading spectra calibrations from file
Details:

Setting the calibration over and over again is a pain in the ass,
frankly. So a new feature has been added that enables the user to use
a GUI to selected an "ADAQ-formateed calibration file". This is a text
file with new columns : column1 == energy, column2 == pulse unit.  The
file should end with an ".acal" extension to identify it to the file
GUI interface. Once the file is set, the calibration is automatically
set upon selection.

Another feature implemented is the ability for the user to select
preset calibration points and see the energy and pulse unit for each
point in the number entry widgets. This is a nice feature for writing
down calibration data.


Former-commit-id: 16b57d13fabe059758301bd8ec5d4b49fa7989b2