Summary: Implemented new feature for setting/plotting calibration line
Details:

The user can now use either the calibration energy/pulse unit to draw
the calibration line on the existing spectrum histogram. This is can
be done either by entering a value in the box and clicking 'enter' or
by using the arrow click buttons. This is extremely useful for setting
calibrations, checking calibrations, or simply seeing where very
specific energies would be in a particular spectrum.

Also, when the user clicks the calibration point from the combo box,
the calibration line will automatically update showing the position of
the calibration lines on the current spectrum. Awesome! This is a very
useful check on the calibration.


Former-commit-id: 207fff530f3614de4709ccd267d1e07cd0964374