Summary: Complete overhaul of AAInterpolation class
Details:

Major overhaul of AAInterpolation with the goal of making it much more
flexible and having much more compact code. All data and TGraphs are
now stored in indexable vectors for access, and response/inverse
function can now by constructed on the fly, which will allow the user
to adjust the light conversion factor on the fly from the interface
(not yet implemented). Tests performed OK.


Former-commit-id: 242d00f329189a7db793253ae26dca6b684977a9