Summary: New env. variable setting for small GUI; odd memory problem hacked
Details:

Setting GUI size via 'small' on the cmd line was a pain since then you
couldn't open files. Since small GUIs are required for a particularly
machine (not switching back-and-forth on the same machine), the user
can now set the environment variable ADAQANALYSIS_SMALL=1 to get a
small GUI automatically.

There is an odd memory problem with if/delete statements in
AAComputation::LoadACRONYMFile(). This is probably due to not
initializing classes in the constructor. Issue doesn't always tickle
bug ... seems OS / architecture dependent since no problems on the
Forge. Commented out if/delete statements so there is small (but
trivial) memory leak ... will look into this as time allows.

Former-commit-id: 3d7135f378e707e34938e35fdc6f6e845f98dbe5