Summary: Beginning branch for massive refactoring of AAInterface
Details:

AAInterface is an enormous ~4500 line class, 4200 lines of which are
in the *.cc file. The purpose of this branch will be to slowly work to
refactor this massive class in to more managable pieces the logically
correspond to the GUI layout. At present, the interface (meaning "the
widgets") will all be built in "AAInterface" (for now); the slots "or
handler" methods, which take about ~2500 lines) will be broken up into
six subclasses, one for each tab and one for the remaining widgets.

This requires very minimal code overhead but will be well worth it in
terms maintainability and ease of updating


Former-commit-id: 8626b5ddd26484a8eb2b6b6afc5e4ef9e33eba10