Started integrating new ADAQ file format
Details:

ADAQAnalysis is now being upgraded to support the new ADAQ file format
provided by the ADAQ libADAQReadout library. This format provides
tremendous improvements in contained data, flexibility,
standardization, etc, over the old file format, which is one of the
last untouched pieces of code from the early days (yeah ... it's
scary...)).

At present, my goal is to attempt to continue supporting the old
"legacy" file format since there are several projects with a LOT of
data in the old format. We'll see how this goes, but currently, the
new ADAQ format is partially integrated and the legacy file format
still works perfectly.


Former-commit-id: dfaebc01a9f5df531983a6b3b2691c98fb7db226