Separated "processing waveforms" and "spectra creation" for much higher efficiency
Details:

Implementing ability to separate the two key actions that go into
making a spectrum: processing of the waveforms (by far the most CPU
and time intensive part); and put analyzed waveform values into a TH1F
to make a spectrum. With this capability, the user only has to process
the waveforms once (for a given set of waveform analysis parameters)
to store the processed waveform values. Then, spectra can be created
from these values very efficienctly. For example, rebinning, new bins
limits, and energy calibration no longer require reprocessing all
waveforms.

There are now two buttons on the spectra tab: a "Process waveforms"
button (that computes waveform values and stored them in a vector for
later use) and a "Create spectrum" button (that uses the vector values
to create spectra). Note that when processing waveforms, a spectra is
automatically created and plotted for convenience.

Initial tests of this new feature were successful. There is
simplification and refactoring that could be done in AAComputation as
well as further testing before release.