Implemented ability to fill PH/PA vectors simultaneously; fix to prevent calibration issues with vectors
Details:

The logic in AAComputation::ProcessSpectraWaveforms() has been changed
to enable calculation of pulse height and area simultaneously
regardless of the chosen value for spectra. This enables the
Spectrum{PH,PA}Vec to be filled, which then allows the user to easily
switch back and forth between PA and PH without reprocessing! A very
nice feature for a very small CPU overhead upfront during processing.

The order of operations has also been changed to ensure that only
uncalibrated PH and PA values are added to the vectors. When spectra
are created FROM the vectors, the values are calibrated
there. Previous check would have caused waveforms to be calibrated
iteratively if multiple processing passes were taken with calibration
enabled